requirements.txt
pyglet
asyncio
aiohttp
pyyaml
